domainname/requirements.txt
asgiref==3.8.1
asttokens==3.0.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
debugpy==1.8.12
decorator==5.1.1
Django==5.1.5
django-extensions==3.2.3
executing==2.2.0
gunicorn==23.0.0
idna==3.10
ipython==8.31.0
jedi==0.19.2
matplotlib-inline==0.1.7
packaging==24.2
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
urllib3==2.3.0
wcwidth==0.2.13
whitenoise==6.8.2
